#5 TUI 実装

+++ emulator/crates/inst/src/def/sra.rs
@@ -8,9 +8,9 @@ pub struct Sra;
 
 impl Inst for Sra {
     fn exec(&self, mut state: State) -> anyhow::Result<State> {
-        let rs1 = state.regs.read(self.rs1)? as i32;
+        let rs1 = state.regs.read(self.rs1)?;
         let rs2 = state.regs.read(self.rs2)?;
-        state.regs.write(self.rd, (rs1 >> rs2) as u32)?;
+        state.regs.write(self.rd, rs1 >> rs2)?;
         state.pc += 6;
         Ok(state)
     }
